pyproject.toml
[tool.poetry]
name = "kingwangjjang-be"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.12"
annotated-types = "0.7.0"
anyio = "4.4.0"
certifi = "2024.7.4"
click = "8.1.7"
colorama = "0.4.6"
dnspython = "2.6.1"
email-validator = "2.2.0"
fastapi = "0.111.1"
fastapi-cli = "0.0.4"
h11 = "0.14.0"
httpcore = "1.0.5"
httptools = "0.6.1"
httpx = "0.27.0"
idna = "3.7"
itsdangerous = "2.2.0"
jinja2 = "3.1.4"
markdown-it-py = "3.0.0"
markupsafe = "2.1.5"
mdurl = "0.1.2"
orjson = "3.10.6"
pydantic = "2.8.2"
pydantic-extra-types = "2.9.0"
pydantic-settings = "2.3.4"
pydantic-core = "2.20.1"
pygments = "2.18.0"
python-dotenv = "1.0.1"
python-multipart = "0.0.9"
pyyaml = "6.0.1"
rich = "13.7.1"
shellingham = "1.5.4"
sniffio = "1.3.1"
starlette = "0.37.2"
typer = "0.12.3"
typing-extensions = "4.12.2"
ujson = "5.10.0"
uvicorn = "0.30.1"
watchfiles = "0.22.0"
websockets = "12.0"
pytest = "^8.2.2"
pytest-asyncio = "^0.23.8"
pymongo = "^4.8.0"
alembic = "^1.13.2"
apscheduler = "^3.10.4"
langchain = "^0.2.9"
ollama = "^0.3.0"
poetry = "^1.8.3"
strawberry-graphql = "^0.236.0"
mongotic = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
